# shared-schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roshetta Pharmacy System",
    "description": "Agentic-AI Pharmacy Operating System for Roshetta (روشتة)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "My Account"
        ],
        "summary": "Get My Profile",
        "description": "Return the full profile of the authenticated user.",
        "operationId": "get_my_profile_api_me_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "RoshettaBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "My Account"
        ],
        "summary": "Update My Profile",
        "description": "Update editable fields on the authenticated user's profile.",
        "operationId": "update_my_profile_api_me_patch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyProfileUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "RoshettaBearer": []
          }
        ]
      }
    },
    "/api/me/photo": {
      "post": {
        "tags": [
          "My Account"
        ],
        "summary": "Upload My Photo",
        "description": "Upload a profile photo; returns the stored URL.",
        "operationId": "upload_my_photo_api_me_photo_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_my_photo_api_me_photo_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "RoshettaBearer": []
          }
        ]
      }
    },
    "/api/me/photo/{filename}": {
      "get": {
        "tags": [
          "My Account"
        ],
        "summary": "Get My Photo",
        "description": "Serve only this account's own generated profile-image key.",
        "operationId": "get_my_photo_api_me_photo__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "RoshettaBearer": []
          }
        ]
      }
    },
    "/api/me/activity": {
      "get": {
        "tags": [
          "My Account"
        ],
        "summary": "Get My Activity",
        "description": "Return this user's sales/expenses/restock summary for the given period.",
        "operationId": "get_my_activity_api_me_activity_get",
        "parameters": [
          {
            "name": "pharmacy_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pharmacy Id"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(day|week|month|all)$",
              "default": "day",
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "RoshettaBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "entry_type": {
            "type": "string",
            "title": "Entry Type"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entry_type",
          "total_amount",
          "payment_method",
          "notes",
          "created_at"
        ],
        "title": "ActivityEntryResponse"
      },
...
      "Body_upload_my_photo_api_me_photo_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_my_photo_api_me_photo_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyActivityResponse": {
        "properties": {
          "period": {
            "type": "string",
            "enum": [
              "day",
              "week",
              "month",
              "all"
            ],
            "title": "Period"
          },
          "pharmacy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pharmacy Id"
          },
          "sales_count": {
            "type": "integer",
            "title": "Sales Count"
          },
          "expenses_count": {
            "type": "integer",
            "title": "Expenses Count"
          },
          "restocks_count": {
            "type": "integer",
            "title": "Restocks Count"
          },
          "total_restock": {
            "type": "number",
            "title": "Total Restock"
          },
          "total_sales": {
            "type": "number",
            "title": "Total Sales"
          },
          "total_expenses": {
            "type": "number",
            "title": "Total Expenses"
          },
          "items_sold": {
            "type": "number",
            "title": "Items Sold"
          },
          "items_restocked": {
            "type": "number",
            "title": "Items Restocked"
          },
          "net": {
            "type": "number",
            "title": "Net"
          },
          "recent_entries": {
            "items": {
              "$ref": "#/components/schemas/ActivityEntryResponse"
            },
            "type": "array",
            "title": "Recent Entries"
          }
        },
        "type": "object",
        "required": [
          "period",
          "pharmacy_id",
          "sales_count",
          "expenses_count",
          "restocks_count",
          "total_restock",
          "total_sales",
          "total_expenses",
          "items_sold",
          "items_restocked",
          "net",
          "recent_entries"
        ],
        "title": "MyActivityResponse"
      },
...
      "MyProfileResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/MyProfileUserResponse"
          },
          "pharmacy_count": {
            "type": "integer",
            "title": "Pharmacy Count"
          },
          "pharmacy_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pharmacy Ids"
          },
          "roles": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "user",
          "pharmacy_count",
          "pharmacy_ids",
          "roles"
        ],
        "title": "MyProfileResponse"
      },
      "MyProfileUpdateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "user": {
            "$ref": "#/components/schemas/MyProfileUserResponse"
          }
        },
        "type": "object",
        "required": [
          "success",
          "user"
        ],
        "title": "MyProfileUpdateResponse"
      },
      "MyProfileUserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "language_pref": {
            "type": "string",
            "title": "Language Pref"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "location",
          "birth_date",
          "language_pref",
          "photo_url",
          "is_active",
          "created_at"
        ],
        "title": "MyProfileUserResponse"
      },
...
      "PhotoUploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "photo_url": {
            "type": "string",
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "success",
          "photo_url"
        ],
        "title": "PhotoUploadResponse"
      },
...
      "ProfileUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "language_pref": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(ar|en)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Pref"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "RoshettaBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}